api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
openai>=1.10.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Auth dependencies
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
sqlalchemy>=2.0.0
